+++ flyfile.js
@@ -1,0 +1,7 @@
+const x = module.exports
+
+x.babel = function * (fly) {
+  fly.source('src/**/*.js')
+    .babel()
+    .target('out')
+}