+++ CHANGELOG.md
@@ -22,4 +22,11 @@
 [1.0.1]
 
   - Minor doc improvements
-   
+
+[1.0.2]
+
+  - Changed `validator` api
+    
+    * validators should return `null` if validation passed and `Error` object if validation failed
+
+  - Validators are now chainable. You can use it like `arrayOf(string).required`